refactored initialize.c -- moved pi 3 specific code to proper places
