+++ system/platforms/arm-rpi3/platforminit.c
@@ -5,7 +5,8 @@
 
 #include <platform.h>
 #include <string.h>
-#include "bcm2837.h"
+#include <bcm2837.h>
+#include <rpi_gpio.h>
 #include "../../../device/uart-pl011/pl011.h"
 #include <mmu.h>
 
@@ -148,6 +149,30 @@ void pl011_init(void)
 	regptr->cr = PL011_CR_RXE | PL011_CR_TXE | PL011_CR_UARTEN;
 
 //	regptr->lcrh |= PL011_LCRH_FEN;
+}
+
+/**
+ * This code is irrelevant to Embedded Xinu and is only used as a means of testing on the RPI 3 boards
+ */
+/* Initialize GPIO pin 16 as an output */
+void led_init(void)
+{
+	volatile struct rpi_gpio_regs *regptr;
+	regptr = (struct rpi_gpio_regs *)(GPIO_REGS_BASE);
+	regptr->gpfsel[1] &= ~(7 << 18);
+	regptr->gpfsel[1] |=  (1 << 18);	
+}
+/* Set GPIO pin 16 to ON */
+void led_on(void)
+{	
+	volatile struct rpi_gpio_regs *regptr = (struct rpi_gpio_regs *)(GPIO_REGS_BASE);
+	regptr->gpset[0] = 1 << 16;
+}
+/* Set GPIO pin 16 to OFF */
+void led_off(void)
+{	
+	volatile struct rpi_gpio_regs *regptr = (struct rpi_gpio_regs *)(GPIO_REGS_BASE);
+	regptr->gpclr[0] = 1 << 16;
 }
 
 /**
